requirements.txt
# ── YouTube API ───────────────────────────────────────────────────────
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# ── Audio processing ──────────────────────────────────────────────────
librosa==0.10.2
pydub==0.25.1
soundfile==0.12.1
pedalboard==0.9.13        # Spotify's audio effects library
numpy==1.26.4
scipy==1.13.1
pyloudnorm==0.1.1         # LUFS normalization (optional but recommended)
numba==0.60.0             # speeds up librosa

# ── Video generation ──────────────────────────────────────────────────
moviepy==1.0.3
Pillow==10.4.0
imageio==2.35.0
imageio-ffmpeg==0.5.1
decorator==4.4.2          # moviepy dependency

# ── Misc ──────────────────────────────────────────────────────────────
requests==2.32.3
tqdm==4.66.5
